memcmp.S: optimize for medium to large sizes
This improved memcmp provides a fast path for compares up to 16 bytes
and then compares 16 bytes at a time, thus optimizing loads from both
sources.  The glibc memcmp microbenchmark retains performance (with an
error of ~1ns) for smaller compare sizes and reduces up to 31% of
execution time for compares up to 4K on the APM Mustang.  On Qualcomm
Falkor this improves to almost 48%, i.e. it is almost 2x improvement
for sizes of 2K and above.

from new lib

Change-Id: I59e6c1e45b33a036f8f77a9931091f986632f356
Signed-off-by: klozz <shaclone707@gmail.com>
Signed-off-by: Thagoo <lohitgowda56@gmail.com>